Refactor agent and shape interfaces for delegation
Introduces IObject and IDelegatingShape interfaces to improve delegation and separation of concerns in agent and shape handling. Refactors IAgent to extend IObject and IDelegatingShape, moving shape delegation logic out of IAgent. Updates MinimalAgent and related classes to align with new interfaces. Also updates copyright years and adds/improves Javadoc comments for clarity.
