+++ gama.core/src/gama/gaml/compilation/IGamaHelper.java
@@ -1,8 +1,8 @@
 /*******************************************************************************************************
  *
- * IGamaHelper.java, in gama.core, is part of the source code of the GAMA modeling and simulation platform (v.1.9.3).
+ * IGamaHelper.java, in gama.core, is part of the source code of the GAMA modeling and simulation platform (v.2025-03).
  *
- * (c) 2007-2024 UMI 209 UMMISCO IRD/SU & Partners (IRIT, MIAT, TLU, CTU)
+ * (c) 2007-2025 UMI 209 UMMISCO IRD/SU & Partners (IRIT, MIAT, ESPACE-DEV, CTU)
  *
  * Visit https://github.com/gama-platform/gama for license information and contacts.
  *
